requirements.txt
# LMS Website Builder
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
